Fixed Animation

+++ include/extensions/phisiks/timer.hpp
@@ -13,8 +13,8 @@
 
 namespace Timer {
 	auto start = std::chrono::system_clock::now ( );
-	float delta_time = 0;
-	float fixed_delta_time = 0;
+	double delta_time = 0;
+	double fixed_delta_time = 0;
 	double current_time = 0;
 	double current_fixed_time = 0;
 
@@ -30,13 +30,13 @@ namespace Timer {
 	void Update ( )
 	{
 		auto t = Get_Time_d ( );
-		delta_time = (float)(t - current_time);
+		delta_time = (t - current_time);
 		current_time = t;
 	}
 	void Fixed_Update ( )
 	{
 		auto t = Get_Time_d ( );
-		fixed_delta_time = (float)(t - current_fixed_time);
+		fixed_delta_time = (t - current_fixed_time);
 		current_fixed_time = t;
 	}
 }
